# Optimize queries comparing a constant value over links