--- a/src/realm/query_expression.hpp
+++ b/src/realm/query_expression.hpp
@@ -451,6 +451,8 @@ template <class oper, class TLeft = Subexpr>
 class UnaryOperator;
 template <class oper, class TLeft = Subexpr>
 class SizeOperator;
+template <class TCond, class T>
+class CompareColumnPathToValue;
 template <class TCond, class T, class TLeft = Subexpr, class TRight = Subexpr>
 class Compare;
 template <bool has_links>
@@ -522,6 +524,15 @@ Query create(L left, const Subexpr2<R>& right)
         }
         // Return query_engine.hpp node
         return q;
+    }
+    else if (column && ((std::numeric_limits<L>::is_integer && std::numeric_limits<R>::is_integer) ||
+                (std::is_same<L, double>::value && std::is_same<R, double>::value) ||
+                (std::is_same<L, float>::value && std::is_same<R, float>::value) ||
+                (std::is_same<L, Timestamp>::value && std::is_same<R, Timestamp>::value) ||
+                (std::is_same<L, StringData>::value && std::is_same<R, StringData>::value) ||
+                (std::is_same<L, BinaryData>::value && std::is_same<R, BinaryData>::value)) &&
+             column->links_exist() && column->only_unary_links()) {
+            return make_expression<CompareColumnPathToValue<Cond, L>>(left, column->column_ndx(), column->get_link_map());
     }
     else
 #endif
@@ -1736,6 +1747,19 @@ struct MakeLinkVector : public LinkMapFunction {
     std::vector<size_t>& m_links;
 };
 
+struct UnaryLinkResult : public LinkMapFunction {
+    UnaryLinkResult()
+        : m_result(realm::not_found)
+    {
+    }
+    bool consume(size_t row_index) override
+    {
+        m_result = row_index;
+        return false; // exit search, only one result ever expected
+    }
+    size_t m_result;
+};
+
 struct CountLinks : public LinkMapFunction {
     bool consume(size_t) override
     {
@@ -1874,7 +1898,15 @@ public:
         return s;
     }
 
-    std::vector<size_t> get_links(size_t index)
+    size_t get_unary_link_or_not_found(size_t index) const
+    {
+        REALM_ASSERT(m_only_unary_links);
+        UnaryLinkResult res;
+        map_links(index, res);
+        return res.m_result;
+    }
+
+    std::vector<size_t> get_links(size_t index) const
     {
         std::vector<size_t> res;
         get_links(index, res);
@@ -1895,7 +1927,7 @@ public:
         return counter.result();
     }
 
-    void map_links(size_t row, LinkMapFunction& lm)
+    void map_links(size_t row, LinkMapFunction& lm) const
     {
         map_links(0, row, lm);
     }
@@ -1924,7 +1956,7 @@ public:
     std::vector<const ColumnBase*> m_link_columns;
 
 private:
-    void map_links(size_t column, size_t row, LinkMapFunction& lm)
+    void map_links(size_t column, size_t row, LinkMapFunction& lm) const
     {
         bool last = (column + 1 == m_link_columns.size());
         ColumnType type = m_link_types[column];
@@ -1973,7 +2005,7 @@ private:
     }
 
 
-    void get_links(size_t row, std::vector<size_t>& result)
+    void get_links(size_t row, std::vector<size_t>& result) const
     {
         MakeLinkVector mlv = MakeLinkVector(result);
         map_links(row, mlv);
@@ -2082,6 +2114,16 @@ public:
     bool links_exist() const
     {
         return m_link_map.m_link_columns.size() > 0;
+    }
+
+    bool only_unary_links() const
+    {
+        return m_link_map.only_unary_links();
+    }
+
+    LinkMap get_link_map() const
+    {
+        return m_link_map;
     }
 
     virtual std::string description(util::serializer::SerialisationState& state) const override
@@ -3789,6 +3831,184 @@ private:
 };
 
 
+template <class TCond, class T>
+class CompareColumnPathToValueBase : public Expression {
+public:
+    CompareColumnPathToValueBase(T left, size_t column, LinkMap link_map)
+    : m_left(std::move(left))
+    , m_column_ndx(column)
+    , m_link_map(std::move(link_map))
+    {
+        REALM_ASSERT(m_link_map.target_table());
+        REALM_ASSERT(m_link_map.links_exist());
+        REALM_ASSERT(m_link_map.only_unary_links());
+        m_column = &m_link_map.target_table()->get_column_base(m_column_ndx);
+        m_right_is_nullable = m_column->is_nullable();
+    }
+
+    bool is_nullable() const noexcept
+    {
+        return m_link_map.base_table()->is_nullable(m_column->get_column_index());
+    }
+
+    const Table* get_base_table() const override
+    {
+        return m_link_map.base_table();
+    }
+
+    void set_base_table(const Table* table) override
+    {
+        if (table != get_base_table()) {
+            m_link_map.set_base_table(table);
+            m_column = &m_link_map.target_table()->get_column_base(m_column_ndx);
+        }
+    }
+
+    void verify_column() const override
+    {
+        // verify links
+        m_link_map.verify_columns();
+        // verify target table
+        const Table* target_table = m_link_map.target_table();
+        if (target_table && m_column_ndx != npos) {
+            target_table->verify_column(m_column_ndx, m_column);
+        }
+    }
+
+    size_t find_first(size_t start, size_t end) const override
+    {
+        size_t match;
+        size_t col = m_column->get_column_index();
+        const Table* target_table = m_link_map.target_table();
+        for (; start < end; ++start) {
+            size_t link_translation_index = m_link_map.get_unary_link_or_not_found(start);
+            if (link_translation_index != realm::not_found) {
+                T rval = target_table->template get<T>(col, link_translation_index);
+                match = cmp(m_left, rval);
+                if (match)
+                    return start;
+            }
+        }
+
+        return not_found; // no match
+    }
+
+    // fixme: test this
+    virtual std::string description(util::serializer::SerialisationState& state) const override
+    {
+        if (std::is_same<TCond, BeginsWith>::value
+            || std::is_same<TCond, BeginsWithIns>::value
+            || std::is_same<TCond, EndsWith>::value
+            || std::is_same<TCond, EndsWithIns>::value
+            || std::is_same<TCond, Contains>::value
+            || std::is_same<TCond, ContainsIns>::value
+            || std::is_same<TCond, Like>::value
+            || std::is_same<TCond, LikeIns>::value) {
+            // these string conditions have the arguments reversed but the order is important
+            // operations ==, and != can be reversed because the produce the same results both ways
+            return util::serializer::print_value(state.describe_columns(m_link_map, m_column_ndx)
+                                                 + " " + TCond::description()
+                                                 + " " + util::serializer::print_value(m_left));
+        }
+        return util::serializer::print_value(state.describe_columns(m_link_map, m_column_ndx) + " " + TCond::description()
+                                             + " " + util::serializer::print_value(m_left));
+    }
+
+    std::unique_ptr<Expression> clone(QueryNodeHandoverPatches* patches) const override
+    {
+        return std::unique_ptr<Expression>(new CompareColumnPathToValueBase(*this, patches));
+    }
+
+//    void apply_handover_patch(QueryNodeHandoverPatches& patches, Group& group) override
+//    {
+//        m_right->apply_handover_patch(patches, group);
+//    }
+
+protected:
+    CompareColumnPathToValueBase(const CompareColumnPathToValueBase& other, QueryNodeHandoverPatches* patches)
+    : m_left(other.m_left)
+    , m_column_ndx(other.m_column_ndx)
+    , m_column(other.m_column)
+    , m_link_map(other.m_link_map, patches)
+    {
+        if (patches && m_column) {
+            m_right_is_nullable = m_column->is_nullable();
+            m_column_ndx = m_column->get_column_index();
+            m_column = nullptr;
+        }
+    }
+    TCond cmp;
+    T m_left;
+    bool m_right_is_nullable;
+    // Column index of payload column of m_table
+    mutable size_t m_column_ndx;
+    const ColumnBase* m_column;
+    LinkMap m_link_map;
+};
+
+
+template <class TCond, class T>
+class CompareColumnPathToValue : public CompareColumnPathToValueBase<TCond, T> {
+public:
+    CompareColumnPathToValue(T left, size_t column, LinkMap link_map)
+    : CompareColumnPathToValueBase<TCond, T>(left, column, link_map)
+    {
+    }
+
+    std::unique_ptr<Expression> clone(QueryNodeHandoverPatches* patches) const override
+    {
+        return std::unique_ptr<Expression>(new CompareColumnPathToValue(*this, patches));
+    }
+
+private:
+    CompareColumnPathToValue(const CompareColumnPathToValue& other, QueryNodeHandoverPatches* patches)
+    : CompareColumnPathToValueBase<TCond, T>(other, patches)
+    {
+    }
+};
+
+
+template <class TCond>
+class CompareColumnPathToValue<TCond, Timestamp> : public CompareColumnPathToValueBase<TCond, Timestamp> {
+public:
+    CompareColumnPathToValue(Timestamp left, size_t column, LinkMap link_map)
+    : CompareColumnPathToValueBase<TCond, Timestamp>(left, column, link_map)
+    {
+    }
+
+    std::unique_ptr<Expression> clone(QueryNodeHandoverPatches* patches) const override
+    {
+        return std::unique_ptr<Expression>(new CompareColumnPathToValue(*this, patches));
+    }
+
+    size_t find_first(size_t start, size_t end) const override
+    {
+        size_t match;
+        size_t col = this->m_column->get_column_index();
+        const Table* target_table = this->m_link_map.target_table();
+        const TimestampColumn* timestamp_col = dynamic_cast<const TimestampColumn*>(this->m_column);
+        for (; start < end; ++start) {
+            size_t link_translation_index = this->m_link_map.get_unary_link_or_not_found(start);
+            if (link_translation_index != realm::not_found) {
+//                timestamp_col->get_secod
+                Timestamp rval = target_table->get<Timestamp>(col, link_translation_index);
+                match = this->cmp(this->m_left, rval, this->m_left.is_null(), this->m_right_is_nullable && rval.is_null());
+                if (match)
+                    return start;
+            }
+        }
+
+        return not_found; // no match
+    }
+
+private:
+    CompareColumnPathToValue(const CompareColumnPathToValue& other, QueryNodeHandoverPatches* patches)
+    : CompareColumnPathToValueBase<TCond, Timestamp>(other, patches)
+    {
+    }
+};
+
+
 template <class TCond, class T, class TLeft, class TRight>
 class Compare : public Expression {
 public:
@@ -3812,8 +4032,7 @@ public:
     }
 
     // Recursively fetch tables of columns in expression tree. Used when user first builds a stand-alone expression
-    // and
-    // binds it to a Query at a later time
+    // and binds it to a Query at a later time
     const Table* get_base_table() const override
     {
         const Table* l = m_left->get_base_table();