… CHANGELOG.md
… test/test_query.cpp