Although not timestamp specific, these changes are for https://jira.mongodb.org/browse/RCORE-4

There are three optimizations here:
1) If comparing to a constant value (applies to `Value<T>`) then do not evaluate this condition every time, only once.
2) If comparing to a link chain property which contains only unary links (ie there are no lists in the path) then we know there will only be one result or null; do not construct and destruct dynamic memory (vectors) for the results.
3) The default storage size of a `Value<T>` constructed by default or for a constant is changed from 8 to 1. Previously, a `Columns` expression would get the next 8 values in the column but only actually use the first one, so we save 7 memory fetches here.

Benchmark results on master:
```
QueryStringOverLinks (MemOnly, EncryptionOff):   min 279.96ms     max 287.08ms     median 284.84ms     avg 283.80ms     stddev   2.46ms
QueryStringOverLinks (MemOnly, EncryptionOn):    min 288.33ms     max 365.90ms     median 297.05ms     avg 303.09ms     stddev  23.55ms
QueryStringOverLinks (Full   , EncryptionOff):   min 284.85ms     max 293.21ms     median 289.10ms     avg 289.08ms     stddev   2.37ms
QueryStringOverLinks (Full   , EncryptionOn):    min 287.63ms     max 295.68ms     median 292.51ms     avg 291.84ms     stddev   2.63ms

QueryTimestampGreaterOverLinks (MemOnly, EncryptionOff):   min  86.87ms     max 100.71ms     median  91.86ms     avg  91.74ms     stddev   5.47ms
QueryTimestampGreaterOverLinks (MemOnly, EncryptionOn):    min  87.30ms     max  90.59ms     median  89.09ms     avg  88.89ms     stddev   1.10ms
QueryTimestampGreaterOverLinks (Full   , EncryptionOff):   min  88.47ms     max 105.86ms     median  92.54ms     avg  93.70ms     stddev   5.57ms
QueryTimestampGreaterOverLinks (Full   , EncryptionOn):    min  86.65ms     max  97.04ms     median  89.25ms     avg  90.18ms     stddev   3.44ms
```

Benchmark results on this branch:
```
QueryStringOverLinks (MemOnly, EncryptionOff):   min 104.39ms     max 139.03ms     median 124.59ms     avg 119.43ms     stddev  10.54ms
QueryStringOverLinks (MemOnly, EncryptionOn):    min 100.67ms     max 111.67ms     median 109.91ms     avg 106.67ms     stddev   4.39ms
QueryStringOverLinks (Full   , EncryptionOff):   min  99.47ms     max 108.40ms     median 106.73ms     avg 104.63ms     stddev   3.25ms
QueryStringOverLinks (Full   , EncryptionOn):    min  98.85ms     max 108.21ms     median 101.07ms     avg 101.89ms     stddev   3.08ms

QueryTimestampGreaterOverLinks (MemOnly, EncryptionOff):   min  41.33ms     max  44.04ms     median  43.08ms     avg  42.57ms     stddev   1.02ms
QueryTimestampGreaterOverLinks (MemOnly, EncryptionOn):    min  40.97ms     max  47.47ms     median  43.59ms     avg  43.29ms     stddev   2.10ms
QueryTimestampGreaterOverLinks (Full   , EncryptionOff):   min  41.10ms     max  56.60ms     median  42.15ms     avg  45.41ms     stddev   6.36ms
QueryTimestampGreaterOverLinks (Full   , EncryptionOn):    min  41.81ms     max  46.98ms     median  42.55ms     avg  43.29ms     stddev   1.99ms
```