@jedelbo 2019-08-30
The PR should be squashed when merged.